requirements.txt
asgiref==3.7.2
awsebcli==3.20.6
Babel==2.12.1
botocore==1.29.99
cement==2.8.2
certifi==2023.5.7
charset-normalizer==2.0.12
colorama==0.4.3
Django==4.2.2
django-admin-honeypot-updated-2021==1.2.0
django-admin-thumbnails==0.2.8
django-ipware==5.0.0
django-phonenumber-field==7.1.0
idna==3.4
jmespath==1.0.1
pathspec==0.10.1
phonenumbers==8.13.13
Pillow==9.5.0
pypiwin32==223
python-dateutil==2.8.2
python-decouple==3.8
pywin32==306
PyYAML==5.4.1
requests==2.26.0
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.4
termcolor==1.1.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
wcwidth==0.1.9
